Make it possible to pick only on ip address
This patch makes it possible to pick out only the first ip address
if the host has more than one, e.g if there is both ipv4 and ipv6
adresses for the host. It should be backwards compatible with the
current implemantation, so that you need to explicitly set the flag
single_ip to get this new behaviour.