Make sure ipv4 address comes first
Prior to d024fb595fe8b28abe5974fc87f633b50dac85d6 ipv4 addresses would
be sorted first. This meant that nagios/naemon running in docker without
ipv6 would be able to ping the host. With the change of sort order this
is no longer the case. So by sorting and reversing the order we make
sure that the same result is always returned AND we get the ipv4 address
first.