Make sorting optional
If you need to have the addresses sorted you can now set
$sort_alphabetically to true.

This is only neccessary if you have more than one ipv4/ipv6 as the
sort order of the dnsLookup function is non determenistic.

Please not that if you do that, you are no longer assured that ipv4
addresses comes first, that they do if you do not sort the array.

In puppet > 5.5 it is possible to have a lambda sort the array, and then
we could have both ipv4 first and deterministic sort order, but ubuntu
22.04 is still on puppet 5.5.